# specs/002-jwt-auth/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Task API with JWT Authentication
  description: |
    RESTful API for managing user tasks with JWT-based authentication.
    All endpoints require a valid JWT token in the Authorization header.
  version: 2.0.0
  contact:
    name: API Support

servers:
  - url: http://localhost:8000
    description: Development server

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued by Better Auth

  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - completed
        - user_id
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique task identifier
          example: 1
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Optional task description
          example: "Milk, bread, eggs"
        completed:
          type: boolean
          description: Task completion status
          example: false
        user_id:
          type: integer
          description: ID of the user who owns this task
          example: 1
        created_at:
          type: string
          format: date-time
          description: Timestamp when task was created
          example: "2026-02-07T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when task was last updated
          example: "2026-02-07T10:30:00Z"

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Optional task description
          example: "Milk, bread, eggs"
        completed:
          type: boolean
          default: false
          description: Initial completion status
          example: false

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Updated task title
          example: "Buy groceries and supplies"
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Updated task description
          example: "Milk, bread, eggs, coffee"
        completed:
          type: boolean
          description: Updated completion status
          example: true

    TaskCompleteRequest:
      type: object
      required:
        - completed
      properties:
        completed:
          type: boolean
          description: New completion status
          example: true

    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Invalid authentication token"

  responses:
    UnauthorizedError:
      description: Authentication failed - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing_token:
              value:
                detail: "Missing authentication token"
            invalid_token:
              value:
                detail: "Invalid authentication token"
            expired_token:
              value:
                detail: "Token has expired"

    ForbiddenError:
      description: Authorization failed - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Not authorized to access this user's resources"

    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Task not found"

    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items:
                        type: string
                    msg:
                      type: string
                    type:
                      type: string

paths:
  /api/tasks:
    get:
      summary: Get all tasks for authenticated user
      description: |
        Retrieves all tasks belonging to the authenticated user.
        Tasks are automatically filtered by the user ID extracted from the JWT token.
      operationId: getTasks
      tags:
        - Tasks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      summary: Create a new task
      description: |
        Creates a new task for the authenticated user.
        The user_id is automatically set from the JWT token.
      operationId: createTask
      tags:
        - Tasks
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /api/tasks/{task_id}:
    get:
      summary: Get a specific task
      description: |
        Retrieves a specific task by ID.
        The authenticated user must own the task.
      operationId: getTask
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      summary: Update a task
      description: |
        Updates an existing task.
        The authenticated user must own the task.
      operationId: updateTask
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Delete a task
      description: |
        Deletes an existing task.
        The authenticated user must own the task.
      operationId: deleteTask
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/tasks/{task_id}/complete:
    patch:
      summary: Mark task as complete or incomplete
      description: |
        Updates the completion status of a task.
        The authenticated user must own the task.
      operationId: updateTaskCompletion
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCompleteRequest'
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

tags:
  - name: Tasks
    description: Task management operations